Fixed an issue with resetting the position property of FallingObject
Set the positionProperty to a new Vector2 object when it is being
edited to prevent the internally stored intialValue from being
edited as well.

Upon construction, Axon Property objects internally store the
initial value of the property in two places: _value and _initialValue.
_value is edited when the property changes and _initialValue is used to
reset the property. Since they are both initially a reference to the same
variable, editing the value for the property also has the potential to
edit the internally stored initialValue, preventing the property from
being reset.