Fixed a few bugs that were preventing ValueGraph from working.
Bug list: (1) When array with a new point added was passed to dataPointsProperty, it determined the given array was not different and did not signal an updated to listeners (this is due to how the given array with new data (in the .set method) had the same memory reference as the internal variable holding the Property's value (._value)). Overwrote the areValuesEqual method to always return false. (2) incrementMaxValue and incrementMaxTime triggered replots before the modelToView transform could be updated, so replotting the data did nothing as the scales weren't changed. (3) getTargetValue on line 96 in ValueGraphNode missing parens. (4) timeScaleProperty and valueScaleProperty on line 59 were missing .get calls. (5) Link on replotGraphProperty (line 138) missing resetPlot call and missing a reset to replotGraphProperty.