Created combusted image view for when incredibly large drag forces create issues.
When objects with a small mass (i.e. ping pong balls and badminton shuttlecocks) are pushed passed their terminal velocity and then drag is enabled in the sim, the drag force becomes so strong that the object's velocity will actually switch direction and the object will continue to accelerate towards positive infinity. To resolve this issue, when the velocity of an object becomes positive it will be put into a 'combusted' state and the sim will become disabled.