Fixed graphing bug that caused time values to stay constant until next replot
Test to replicate the issue before commit: launch sim, enable graphs, switch to Badminton Shuttlecock, press play and then immediately enable drag. New points on the acceleration graph will have a constant X value until the graph's xscale changes. Caused by a disconnect between the timeScale and the xScale in the ValueGraphNode's modelViewTransform. Switching to the Badminton Shuttlecock resets the ValueGraphModel, setting the timeScale to zero, however the xScale in the modelViewTransform is not updated as it should be. This is why values will continue to plot normally. When drag is enabled, it triggers a valueScale change, which changes the originLocPercent, which triggers the modelViewTranform to be recreated. Then zero timeScale is then used for the xScale in the modelViewTransform, causing all xvalues to plot to the same point (x=50 in view coordinates). Set the initial values for the time and value scale to reflect the timeLength and valueLength, instead of just being zero.

Next commit will have the modelViewTransform automatically update its x and y scales to match the timeScale and valueScale.